### 2023-08-03 #5733 @editorialbot
[PRE REVIEW]: WPSS, a web panel sample service
**Submitting author:** @tvi-cdsp (<a href="http://orcid.org/0000-0002-7249-4228">Tom Villette</a>)
**Repository:** https://github.com/CDSP-SCPO/WPSS-for-ESS-webpanel
**Branch with paper.md** (empty if default branch): joss
**Version:** 2023-06-21
