### 2023-10-01 #5733 @arfon
:wave: @ajstewartlang – ☝️ this list of the past most similar papers published in JOSS might be a good source of possible reviewers. Please ask past _authors_ first before approaching past reviewers.
