### 2023-10-03 #5733 @pmkruyen
👋 Thank you for reaching out, the software seems very promising to me. I'm however unsure if I have time the next few weeks to review the paper & software. Could you please try to find other reviewers first and else come back to me in about two weeks?

Meanwhile, we use Qualtrics at our university. How can I install / integrate this in Qualtrics? For me the instructions on Github were not that clear. Should I consider this as a plugin that I can upload in Qualtrics?
