### 2023-10-09 #5733 @tvi-cdsp
Hello, this is not a plugin to upload in Qualtrics but a webserver communicating with Qualtrics through it's API.

Some configurations have to be done in Qualtrics to allow the webserver to make full use of Qualtrics api through (like creating a dedicated account & a 'technical' survey / library / directory...)
