### 2023-11-11 #6022 @joaojcorreia
Below you will find some of my notes on the article and on the repository documentation. I am currently without a functioning Qualtrics account, something I hope my institution's IT department will sort out in the next couple of weeks. Once I have it, I will move forward with installing and testing WPSS.

____________________________________________________________

**Notes on the article:**

- General note: The authors often use a conditional tone when discussing the potential use cases for WPSS. Given the tool's significant value, a more assertive tone is recommended. Emphasizing the value of WPSS will enhance the overall communication.
- Summary: Well-written and effectively conveys the intended message. However, I suggest placing greater emphasis on the practical aspects of WPSS.
- Statement of need: While all the components are present, I suggest reorganizing the text to better highlight the value and necessity of WPSS:

1. Emphasize the significance of cross-national web surveys.
2. Highlight the challenges associated with managing such surveys.
3. Demonstrate how WPSS effectively addresses these challenges, providing added value to researchers engaged in cross-national web surveys.
4. Clearly articulate the reasons for choosing a tool integrated with Qualtrics.

**Notes on the documentation:**

- The statement of need currently lacks clarity and depth. To enhance its effectiveness, consider further development, including the incorporation of a couple of brief, generic use cases directly in the GitHub README. While detailed presentations of various use cases are available on the project website, having concise examples in the README would provide potential users with immediate clarity on the tool's relevance and utility.
