### 2024-01-22 #6022 @ajstewartlang
:wave: @joaojcorreia @pmkruyen I've been trying to contact @tvi-cdsp via GH and via email (using a couple of different addresses) asking them to reply to the queries above. So far I haven't had any response. I'll give it another few weeks - if I don't hear back soon, I'll have to reject this submission unfortunately.
