### 2024-01-26 #6022 @pmkruyen
Dear all,

I understand the situation, but because of change in positions (and
employer), I can no longer access Qualtrics. I have to withdraw from the
review. Hope you find another reviewer. I am very sorry for this.

Kind regards,
Peter

On Tue, Jan 23, 2024 at 5:38 PM Tom Villette ***@***.***>
wrote:

> Hello,
> i apologise for the long delay in our response,
> The team member in charge of monitoring the GH repo left our team and no
> one else here was watching it, also my own work Github account was set up
> incorrectly and i did not receive any notifications until the email to
> info.cdsp...
> All of this is hopefully fixed now, and @pmkruyen
> <https://github.com/pmkruyen> query should also be answered in our github
> issues.
>
> Again, sorry for the delay, the holiday season did not help either here.
>
