### 2024-02-29 0a84391
Typo fixes in paper
There may be further changes that would help readability, but I will address those in the JOSS review.
