+++ paper.md
@@ -58,38 +58,38 @@ date: 24 July 2023
 
 # Summary
 
-WPSS stands for Web Panel Sample Service. It is a web application paired to Qualtrics survey platform to meet the specific needs of cross-national longitudinal web surveys.
-WPSS enables a centralized management of survey fielding tasks based on user roles (publishing, sending invites and reminders), and a decentralized and privacy-compliant handling of panelist data. The service was used by the ESS ERIC (European Social Survey, supported by a European Research Infrastructure Consortium or ERIC) from 2021 to 2023, for the CRONOS2 study ("The CROss-National Online Survey, second edition"), but could be used in similar use cases. An extensive user documentation for the CRONOS2 use case is available online at https://cdsp-scpo.github.io/wpss-doc/.
+Web Panel Sample Service (WPSS) is a web application paired to the Qualtrics survey platform to meet the specific needs of cross-national longitudinal web surveys.
+WPSS enables a centralized management of survey fielding tasks based on user roles (publishing, sending invites and reminders), and a decentralized and privacy-compliant handling of panelist data. The service was used by the ESS ERIC (European Social Survey, supported by a European Research Infrastructure Consortium or ERIC) from 2021 to 2023, for the CRONOS2 study ("The CROss-National Online Survey, second edition"), but could be used in similar use cases. An extensive user documentation for the CRONOS2 use case is available online at <https://cdsp-scpo.github.io/wpss-doc/>
 
 # Statement of need
 
 To meet the need of cross-national web surveys,
-Researchers are in great demand of data collected through surveys, especially for the Social Sciences and Humanities. Online or web surveys are a way to start collecting data in an efficient way, regarding costs and time, compared for example to interviews. Online surveys are also conduced within harmonised and comparative studies such as SHARE, the Survey of Health, Ageing and Retirement in Europe, GGP, the Generations and Gender Programme, EVS, the European Values Study, or Pew Research Center international surveys.
+researchers are in great demand of data collected through surveys, especially for the Social Sciences and Humanities. Online or web surveys are a way to start collecting data in an efficient way, regarding costs and time, compared for example to interviews. Online surveys are also conduced within harmonised and comparative studies such as SHARE, the Survey of Health, Ageing and Retirement in Europe, GGP, the Generations and Gender Programme, EVS, the European Values Study, or Pew Research Center international surveys.
 
 The global service provided should allow to manage respondents samples (importing, managing and exporting respondents' contact data), design and translate questionnaires, draft and translate messages, distribute questionnaires to respondents using messages (both in relevant language version). Respondents should be able to connect to a web application to access questionnaires alternatively, a so-called "panelist portal".
 
 A share of the features above can be provided by existing survey platform, the term "survey platform" refers to the software application used to design web questionnaires and publish them to a set of panellists that can use the platform to read the questions and then submit their answers online. There is no shortage of existing web survey platforms, generally, along with user-friendly questionnaire design tools, they allow study coordinators to manage lists of contacts to which surveys may be distributed through different communication channels, often emails and Short Text Messages. We selected Qualtrics, to be integrated using the Qualtrics API to a custom web application providing additional features.
 
[…]
