### 2024-05-09 #6022 @ajstewartlang
> Hello, i apologise for the long delay in our response, The team member in charge of monitoring the GH repo left our team and no one else here was watching it, also my own work Github account was set up incorrectly and i did not receive any notifications until the email to info.cdsp... All of this is hopefully fixed now, and @pmkruyen query should also be answered in our github issues.
>
> Again, sorry for the delay, the holiday season did not help either here.

:wave: @tvi-cdsp Can I check that the issues that @joaojcorreia mentioned above have been addressed?
