### 2024-06-27 CDSP-SCPO/WPSS-for-ESS-webpanel#3 @ajstewartlang
JOSS review response needed
Can one of the WPSS authors respond to the JOSS review please?

https://github.com/openjournals/joss-reviews/issues/6022#issuecomment-2102388574

thanks

Andrew
