### 2024-08-19 CDSP-SCPO/WPSS-for-ESS-webpanel#3 @ajstewartlang
@gmi-cdsp @tvi-cdsp tagging you both to see if you can help with the above. A response by the end of August would be appreciated.
