### 2024-08-27 fc61760
Update paper.md according to joaojcorreia comment
