+++ paper.md
@@ -58,22 +58,24 @@ date: 24 July 2023
 
 # Summary
 
-Web Panel Sample Service (WPSS) is a web application paired to the Qualtrics survey platform to meet the specific needs of cross-national longitudinal web surveys.
-WPSS enables a centralized management of survey fielding tasks based on user roles (publishing, sending invites and reminders), and a decentralized and privacy-compliant handling of panelist data. The service was used by the ESS ERIC (European Social Survey, supported by a European Research Infrastructure Consortium or ERIC) from 2021 to 2023, for the CRONOS2 study ("The CROss-National Online Survey, second edition"), but could be used in similar use cases. An extensive user documentation for the CRONOS2 use case is available online at <https://cdsp-scpo.github.io/wpss-doc/>
+Web Panel Sample Service (WPSS) is a web application integrated with the Qualtrics survey platform, designed specifically to address the needs of cross-national longitudinal web surveys. WPSS enables centralized management of survey fieldwork tasks based on user roles (publishing, sending invites and reminders), while ensuring decentralized and privacy-compliant handling of panelist data. This service was successfully used by the ESS ERIC (European Social Survey, supported by a European Research Infrastructure Consortium or ERIC) from 2021 to 2023 for the CRONOS2 study ("The CROss-National Online Survey, second edition") and is equally applicable to similar use cases. Comprehensive user documentation for the CRONOS2 use case is available online at <https://cdsp-scpo.github.io/wpss-doc/>
 
 # Statement of need
 
-To meet the need of cross-national web surveys,
-researchers are in great demand of data collected through surveys, especially for the Social Sciences and Humanities. Online or web surveys are a way to start collecting data in an efficient way, regarding costs and time, compared for example to interviews. Online surveys are also conduced within harmonised and comparative studies such as SHARE, the Survey of Health, Ageing and Retirement in Europe, GGP, the Generations and Gender Programme, EVS, the European Values Study, or Pew Research Center international surveys.
+Cross-national web surveys are crucial for research, particularly in the Social Sciences and Humanities. Collecting data through online surveys is an efficient method, both in terms of cost and time, especially when compared to alternatives like interviews. These surveys are often conducted within harmonized and comparative studies, such as SHARE (Survey of Health, Ageing and Retirement in Europe), GGP (Generations and Gender Programme), EVS (European Values Study), or Pew Research Center international surveys.
 
-The global service provided should allow to manage respondents samples (importing, managing and exporting respondents' contact data), design and translate questionnaires, draft and translate messages, distribute questionnaires to respondents using messages (both in relevant language version). Respondents should be able to connect to a web application to access questionnaires alternatively, a so-called "panelist portal".
+However, managing such surveys presents several challenges:
 
-A share of the features above can be provided by existing survey platform, the term "survey platform" refers to the software application used to design web questionnaires and publish them to a set of panellists that can use the platform to read the questions and then submit their answers online. There is no shortage of existing web survey platforms, generally, along with user-friendly questionnaire design tools, they allow study coordinators to manage lists of contacts to which surveys may be distributed through different communication channels, often emails and Short Text Messages. We selected Qualtrics, to be integrated using the Qualtrics API to a custom web application providing additional features.
+    Respondent Sample Management: Importing, managing, and exporting respondents' contact data.
+    Questionnaire Design and Translation: Designing and translating questionnaires, as well as drafting and translating messages.
+    Questionnaire Distribution: Distributing questionnaires to respondents using messages, ensuring they receive the correct language version.
+    Panelist Access: Respondents should have access to a web application, known as the "panelist portal," where they can complete the questionnaires.
 
[…]
