### 2024-08-27 da47e70
Update paper.md typo
