### 2024-08-27 #6022 @Tepau
👋  @ajstewartlang. Sorry for the delay, my two colleagues mentioned above are no longer part of the team, so I'm taking over this subject. I've just pushed the modifications requested by @joaojcorreia , and I hope you'll find them satisfactory. I'm now available to finalize the publication of this software paper.
